Remove unneeded module for S3 and GCS
Consolidate s3_handler and gcs_handler in spack.util
