+++ lib/spack/spack/util/gcs.py
@@ -10,6 +10,10 @@ integrate GCS Blob storage with spack buildcache.
 
 import os
 import sys
+import urllib.parse
+import urllib.response
+from urllib.error import URLError
+from urllib.request import BaseHandler
 
 import llnl.util.tty as tty
 
@@ -222,3 +226,23 @@ class GCSBlob:
         }
 
         return headers
+
+
+def gcs_open(req, *args, **kwargs):
+    """Open a reader stream to a blob object on GCS"""
+    import spack.util.gcs as gcs_util
+
+    url = urllib.parse.urlparse(req.get_full_url())
+    gcsblob = gcs_util.GCSBlob(url)
+
+    if not gcsblob.exists():
+        raise URLError("GCS blob {0} does not exist".format(gcsblob.blob_path))
+    stream = gcsblob.get_blob_byte_stream()
+    headers = gcsblob.get_blob_headers()
+
+    return urllib.response.addinfourl(stream, headers, url)
+
+
+class GCSHandler(BaseHandler):
+    def gs_open(self, req):
+        return gcs_open(req)
